lito/pyproject.toml
[tool.poetry]
name = "sri_papeles_trabajo"
version = "0.1.0"
description = ""
authors = ["Jimmy Gonzalez <jagonzalej@sri.com.ec>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.2"
hypercorn = "^0.16.0"
flask-session = "^0.6.0"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
toml = "^0.10.2"
numpy = "^1.26.4"
joblib = "^1.3.2"
nh3 = "^0.2.15"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
python-calamine = "^0.2.0"
packaging = "^23.2"
flask-compress = "^1.14"
flask-cors = "^4.0.0"
flask-bcrypt = "^1.0.1"
flask-wtf = "^1.2.1"
flask-talisman = "^1.1.0"
flask-psycopg2 = "^1.3"
pyrfc3339 = "^1.1"
asgiref = "^3.7.2"
xlrd = "^2.0.1"
pandarallel = "^1.6.5"
oracledb = "^2.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
